D/os will not support suspending threads
Horribly kludgy implementation of suspending a thread was bothering
me to the point of realizing that this indeed is a very bad idea.
You simply cannot expect to be able to resume a thread after
suspending it in a random moment. What if the thread was waiting
for some event that got lost. Allowing for programming this way
is (intentionally) leaving people for trouble.
You need to program the thread to be fine with suspending at any
random moment. But once you are coding it in a special way, why
not just code it for some cooperative way of suspending?
Simple Lock acquired in main loop is enough...

I hope that explains why D/os will do it the Linux way.
We've learned some lessons since BeOS was concieved.